Extracted ParentableModel
