+++ src/main/java/com/packagename/myapp/models/BaseModel.java
@@ -1,11 +1,8 @@
 package com.packagename.myapp.models;
 
-import com.google.common.collect.Lists;
 import com.packagename.myapp.Application;
-import com.packagename.myapp.models.annotations.Parent;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Text;
-import com.vaadin.flow.component.combobox.ComboBox;
 import com.vaadin.flow.component.textfield.NumberField;
 import com.vaadin.flow.component.textfield.TextField;
 import org.apache.commons.lang3.StringUtils;
@@ -15,17 +12,13 @@ import org.reflections.ReflectionUtils;
 import org.springframework.data.repository.CrudRepository;
 
 import javax.persistence.Table;
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.StreamSupport;
 
-public abstract class BaseModel {
+public abstract class BaseModel extends ParentableModel {
     private final static Logger logger = LogManager.getLogger(BaseModel.class);
 
     public abstract int getId();
@@ -33,12 +26,6 @@ public abstract class BaseModel {
     public abstract String getName();
 
     public abstract void setName(String name);
-
-    // TODO: 09-Oct-20 Use @Parent for getting parent
-    public abstract BaseModel getParent();
-
-    // TODO: 12-Oct-20 Setter using reflection
-    public abstract void setParent(BaseModel parent);
 
 //    public void setParent(BaseModel parent) {
 //        getParentSetter().ifPresent(method -> {
@@ -49,8 +36,6 @@ public abstract class BaseModel {
 //            }
 //        });
 //    }
-
-    public abstract List<BaseModel> getChildren();
 
     @Override
     public boolean equals(Object obj) {
@@ -95,59 +80,6 @@ public abstract class BaseModel {
         return StringUtils.capitalize(getEntityTableName());
     }
 
-    public List<BaseModel> getParentsTree() {
-        ArrayList<BaseModel> parentsTree = new ArrayList<>();
-
-        BaseModel currentParent = getParent();
-        while (currentParent != null) {
-            parentsTree.add(currentParent);
-
-            currentParent = currentParent.getParent();
-        }
-
-        return parentsTree;
-    }
-
-    public List<ComboBox<BaseModel>> getParentTreeCombobox() {
-        ArrayList<ComboBox<BaseModel>> fields = new ArrayList<>();
-
-        BaseModel current = getParentNewInstance();
-        while (current != null) {
-            String propertyName = current.getEntityTableNameCapitalized();
-            CrudRepository<? extends BaseModel, Integer> repository = current.getRepository();
-
-            ComboBox<BaseModel> field = new ComboBox<>(propertyName);
-            field.setItemLabelGenerator(BaseModel::getName);
-            field.setItems(Lists.newArrayList(repository.findAll()));
-            field.setPlaceholder(current.getEntityTableNameCapitalized());
-
-            fields.add(field);
-
-            current = current.getParentNewInstance();
-        }
-
-        return fields;
-    }
-
-    public BaseModel getParentNewInstance() {
-        Optional<? extends Class<?>> parentClass = getParentClass();
-
-        if (parentClass.isPresent()) {
-            try {
-                return (BaseModel) parentClass.get().getDeclaredConstructor().newInstance();
-            } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
-                logger.error("Failed to create new parent instance: " + parentClass.get().getSimpleName(), e);
-            }
-        }
-        return null;
-    }
-
-    public Optional<? extends Class<?>> getParentClass() {
-        Optional<Field> parentField = getParentField();
-
-        return parentField.isPresent() ? Optional.of(parentField.get().getType()) : Optional.empty();
-    }
-
     public List<Component> getPropertiesField() {
         List<Class<?>> acceptedReturnType = Arrays.asList(String.class, Integer.class, BaseModel.class);
 
@@ -176,7 +108,7 @@ public abstract class BaseModel {
                     }
 
                     fields.add(new Text(propertyName + ": Undefined type"));
-                    logger.warn("Not a valid field type: "+propertyName);
+                    logger.warn("Not a valid field type: " + propertyName);
                 });
 
         return fields;
@@ -191,23 +123,7 @@ public abstract class BaseModel {
         return (CrudRepository<T, Integer>) Application.context.getBean(getRepositoryName());
     }
 
-    public boolean hasParent() {
-        return getParentField().isPresent();
-    }
-
-    private Optional<Field> getParentField() {
-        return ReflectionUtils.getAllFields(this.getClass(), ReflectionUtils.withAnnotation(Parent.class)).stream().findFirst();
-    }
-
-    private Optional<Method> getParentSetter() {
-        return ReflectionUtils.getAllMethods(this.getClass(),
-                ReflectionUtils.withModifier(Modifier.PUBLIC),
-                ReflectionUtils.withName("set" + getParentNewInstance().getEntityTableNameCapitalized()))
-                .stream().findFirst();
-    }
-
-    public boolean checkNameAvailability(String name){
+    public boolean existsByName(String name) {
         return StreamSupport.stream(this.getRepository().findAll().spliterator(), false).noneMatch(t -> t.getName().equals(name));
     }
-
 }
